[REFACTOR] Organized into new folders
