+++ src/react-app/Components/Edges/Custom.tsx
@@ -2,24 +2,22 @@ import {
     BaseEdge, 
     EdgeLabelRenderer,
     getBezierPath,
-    // useReactFlow,
+    // useReactFlow, 
     type Edge,
     type EdgeProps,
 } from '@xyflow/react';
 
-import '../index.css';
-import useStore from '../store';
+import '../../styles/index.css';
+import useStore from '../../stores/store';
 
-export type RevMAEdgeType = Edge<{ 
+export type CustomEdgeType = Edge<{ 
     label: string; 
     // toggleDrawer: (id: string) => void;
     rate_law: string;
     rate_type: string; 
-}, 'rev_mass_action'>;
+}, 'custom'>;
 
-
-
-export default function ReversibleMassActionEdge({ 
+export default function CustomEdge({ 
     id, 
     // sourceX, 
     // sourceY, 
@@ -30,9 +28,7 @@ export default function ReversibleMassActionEdge({
     selected,
     markerEnd,
     data,
-}: EdgeProps<RevMAEdgeType>) {
-
-    // const reactantIDs = useStore(store => store.reactions.find(r => r.id === id)?.participants.filter(p => p.role === 'reactant').map(p => p.id)) ?? [];
+}: EdgeProps<CustomEdgeType>) {
     const reactantIDs = useStore(store => store.reactions.find(r => r.id === id)?.participants)?.filter(p => p.role === 'reactant').map(p => p.id) ?? [];
     const productIDs = useStore(store => store.reactions.find(r => r.id === id)?.participants)?.filter(p => p.role === 'product').map(p => p.id) ?? [];
     
@@ -98,7 +94,6 @@ export default function ReversibleMassActionEdge({
     const setEdgeHovering = useStore((store) => store.setEdgeHovering);
     const setEdgeHoverID = useStore((store) => store.setEdgeHoverID);
 
-
     // const onToggle = () => {
     //     data?.toggleDrawer(id);
     // }
@@ -140,7 +135,7 @@ export default function ReversibleMassActionEdge({
             <polyline
               className="arrowclosed"
               style={{
-                strokeWidth: '2px',
+                strokeWidth: 1,
                 stroke: '#747bff',
                 fill: '#747bff',
               }}
@@ -151,6 +146,8 @@ export default function ReversibleMassActionEdge({
           </marker>
         </defs>
       </svg>
+
+
 
 
             {/* Render all forward paths */}
@@ -189,7 +186,6 @@ export default function ReversibleMassActionEdge({
                 ))
             }
 
-
             <EdgeLabelRenderer>
                 <button 
                 onClick={onToggle}
@@ -207,21 +203,17 @@ export default function ReversibleMassActionEdge({
                 onMouseLeave={() => onLeave()}
                 > {data?.label ?? 'Default Label'} </button>
             </EdgeLabelRenderer>
-
         </>
-
-
-        
 
     );
 }
 
-// Specific drawer info for Reversible Mass Action edges. Shows reactants, products, and rate laws.
-export function ReversibleMassActionDrawerInfo() {
+// Specific drawer info for Custom edges.
+export function CustomDrawerInfo() {
     return (
         <>
         
-        {/* <p> REVERSIBLE MASS ACTION TEST </p> */}
+        {/* <p> CUSTOM TEST {edgeID} </p> */}
 
         </>
     );
