+++ src/react-app/Components/Edges/MassAction.tsx
@@ -7,8 +7,8 @@ import {
     type EdgeProps,
 } from '@xyflow/react';
 
-import '../index.css';
-import useStore from '../store';
+import '../../styles/index.css';
+import useStore from '../../stores/store';
 
 export type MAEdgeType = Edge<{ 
     label: string;
