+++ src/react-app/styles/index.css
@@ -1,3 +1,7 @@
+/* This is the base style sheet. 
+I made some mistakes when starting out and put all of the styles in one file (this one), 
+so we're gradually moving to putting all the styles in their correct style sheets! */
+
 :root {
 	font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
 	line-height: 1.5;
@@ -17,14 +21,14 @@
 
 @font-face {
 	font-family: Helios_Extended;
-	src: url(Helios_Extended.otf);
+	src: url(../assets/Helios_Extended.otf);
 	font-weight: normal;
 	font-style: normal;
 }
 
 @font-face {
 	font-family: Helios_Extended;
-	src: url(heliosext_bold.otf);
+	src: url(../assets/heliosext_bold.otf);
 	font-weight: 900;
 	font-style: normal;
 }
